partC_facial_analysis/requirements.txt
azure-functions==1.23.0
azure-core==1.34.0
azure-identity==1.23.0
requests==2.32.4
typing-extensions==4.14.0
pytest==8.4.1
opencv-python==4.10.0.84
numpy==1.26.4
python-dotenv==1.1.1functions==1.23.0
azure-core==1.34.0
azure-identity==1.23.0
requests==2.32.4
typing_extensions==4.14.0
pytest==8.4.1
